+++ R/do_SCExpressionHeatmap.R
@@ -19,6 +19,7 @@ do_SCExpressionHeatmap <- function(sample,
                                    features.order = NULL,
                                    metadata = NULL,
                                    metadata.colors = NULL,
+                                   colorblind = FALSE,
                                    subsample = NA,
                                    cluster = TRUE,
                                    interpolate = FALSE,
@@ -83,7 +84,8 @@ do_SCExpressionHeatmap <- function(sample,
                        "legend.byrow" = legend.byrow,
                        "use_viridis" = use_viridis,
                        "cluster" = cluster,
-                       "interpolate" = interpolate)
+                       "interpolate" = interpolate,
+                       "colorblind" = colorblind)
   check_type(parameters = logical_list, required_type = "logical", test_function = is.logical)
   # Check numeric parameters.
   numeric_list <- list("font.size" = font.size,
@@ -369,7 +371,7 @@ do_SCExpressionHeatmap <- function(sample,
         colors.use <- metadata.colors[[name]]
       } else {
         names.use <- if(is.factor(sample@meta.data[, name])){levels(sample@meta.data[, name])} else {sort(unique(sample@meta.data[, name]))}
-        colors.use <- generate_color_scale(names_use = names.use)
+        colors.use <- generate_color_scale(names_use = names.use, colorblind = colorblind)
       }
       p <- plot_data %>%
            ggplot2::ggplot(mapping = ggplot2::aes(x = .data$cell,